test: line 설계변경에 맞추어 테스트 수정
