+++ src/test/java/domain/LinesTest.java
@@ -1,9 +1,10 @@
 package domain;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
+import domain.lines.Lines;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class LinesTest {
     @Test
@@ -11,8 +12,8 @@ class LinesTest {
     void createLines() {
         int height = 5;
         int personCount = 4;
-        
         Lines lines = new Lines(height, personCount);
+
         assertEquals(height, lines.getLines().size());
     }
 }
